Fixed warnings

+++ src/runtime/Graph.cpp
@@ -199,7 +199,7 @@ void GraphMetadata::map_task_inputs_outputs()
         task_uid_hash_[task->uid()] = task;
     }
     set<unsigned long> output_flushes;
-    for(int index=0; index<tasks.size(); index++) {
+    for(uint32_t index=0; index<(uint32_t)tasks.size(); index++) {
         Task *task = tasks[index];
         unsigned long uid = task->uid();
         vector<unsigned long> input_mems;
@@ -222,7 +222,7 @@ void GraphMetadata::map_task_inputs_outputs()
                       mem_index_hash_[uid] = cmd->mem(); 
                   //_info(" mid:%lu is added", uid);
                   break;
-              case IRIS_CMD_MEM_FLUSH:    
+              case IRIS_CMD_MEM_FLUSH:    // Fallthrough case
                   output_flushes.insert(cmd->mem()->uid());
               case IRIS_CMD_D2H:          
                   uid = cmd->mem()->uid();       
@@ -313,7 +313,7 @@ void GraphMetadata::get_dependency_matrix(int8_t *dep_matrix, bool adj_matrix) {
       dep_adj_list_   = (int8_t *)calloc(ntasks*(ntasks+1), sizeof(int8_t));
       dep_matrix = dep_adj_list_;
   }
-  for(int t=0; t<tasks.size(); t++) {
+  for(uint32_t t=0; t<(uint32_t)tasks.size(); t++) {
       Task *task = tasks[t];
       for(int i=0; i<task->ndepends(); i++) {
           Task *dtask = task->depend(i);
@@ -347,7 +347,7 @@ void GraphMetadata::get_2d_comm_adj_matrix(size_t *comm_task_adj_matrix)
         comm_task_adj_matrix_ = (size_t *)calloc(ntasks*ntasks, sizeof(size_t));
         comm_task_adj_matrix = comm_task_adj_matrix_;
     }
-    for(int index=0; index<tasks.size(); index++) {
+    for(uint32_t index=0; index<(uint32_t)tasks.size(); index++) {
         Task *each_task = tasks[index];
         vector<unsigned long> & lst1_v = task_inputs_map_[each_task->uid()];
         set<unsigned long> lst1(lst1_v.begin(), lst1_v.end());
@@ -423,7 +423,7 @@ void GraphMetadata::get_3d_comm_data()
     vector<Task *> & tasks = graph_->tasks_list();
     int ntasks = tasks.size()+1;
     vector<CommData3D> results;
-    for(int index=0; index<tasks.size(); index++) {
+    for(uint32_t index=0; index<(uint32_t)tasks.size(); index++) {
         Task *each_task = tasks[index];
         vector<unsigned long> & lst1_v = task_inputs_map_[each_task->uid()];
         set<unsigned long> lst1(lst1_v.begin(), lst1_v.end());
@@ -459,7 +459,7 @@ void GraphMetadata::get_3d_comm_data()
                 BaseMem *mem = mem_index_hash_[mid];
                 //printf("Common mem:%lu mid:%lu\n", mem->uid(), mid);
                 size += mem->size();
-                results.push_back({d_index+1, index+1, mid, mem->size()});
+                results.push_back({(uint32_t)d_index+1, (uint32_t)index+1, (uint32_t)mid, mem->size()});
                 all_covered_mem.insert(mid);
             }
             //_info("partial updating %lu-> %lu size:%lu\n", index+1, d_index+1, size);
@@ -474,7 +474,7 @@ void GraphMetadata::get_3d_comm_data()
                     all_covered_mem.insert(dmem_index);
                     all_covered_mem.insert(mid);
                     size += mem->size();
-                    results.push_back({d_index+1, index+1, mid, mem->size()});
+                    results.push_back({(uint32_t)d_index+1, (uint32_t)index+1, (uint32_t)mid, mem->size()});
                 }
             }
             //_info("updating %lu-> %lu size:%lu\n", index+1, d_index+1, size);
@@ -485,7 +485,7 @@ void GraphMetadata::get_3d_comm_data()
             if (all_covered_mem.find(mid) == all_covered_mem.end()) {
                 BaseMem *mem = mem_index_hash_[mid];
                 size += mem->size();
-                results.push_back({0, index+1, mid, mem->size()});
+                results.push_back({0, (uint32_t)index+1, (uint32_t)mid, mem->size()});
             }
         }
     }
